Clean/requirements.txt
torch
loguru
numpy
pandas
tqdm
transformers
boto3
fastparquet
pyarrow
python-dotenv